requirements.txt
numpy
pytrec_eval
tensorflow
transformers
tf-models-official
torch
faiss-cpu
faiss-gpu
